Remove lazy import
